// apps/console/src/pages/organizations/registries/__generated__/RegistriesPageFragment.graphql.ts
export type NonconformityRegistryStatus = "CLOSED" | "IN_PROGRESS" | "OPEN";
import { FragmentRefs } from "relay-runtime";
export type RegistriesPageFragment$data = {
  readonly id: string;
  readonly nonconformityRegistries: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly audit: {
          readonly framework: {
            readonly id: string;
            readonly name: string;
          };
          readonly id: string;
        };
        readonly correctiveAction: string | null | undefined;
        readonly createdAt: any;
        readonly dateIdentified: any | null | undefined;
        readonly description: string | null | undefined;
        readonly dueDate: any | null | undefined;
        readonly effectivenessCheck: string | null | undefined;
        readonly id: string;
        readonly owner: {
          readonly fullName: string;
          readonly id: string;
        };
        readonly referenceId: string;
        readonly rootCause: string;
        readonly status: NonconformityRegistryStatus;
        readonly updatedAt: any;
      };
    }>;
    readonly pageInfo: {
      readonly endCursor: any | null | undefined;
      readonly hasNextPage: boolean;
    };
    readonly totalCount: number;
  };
  readonly " $fragmentType": "RegistriesPageFragment";
};
export type RegistriesPageFragment$key = {
  readonly " $data"?: RegistriesPageFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"RegistriesPageFragment">;
};